Enforce using the compiler daemon if possible
This adds general support for using the compile daemon
over dropping down to the cli compiler if possible.
Most of the changes are about removing the assumption
that the current JVM is the worker/compiler default.

Includes a bug fix to properly handle the release flag
if both release and targetCompatibility are set without
using the new toolchain support.